Move sample_next to MetropolisSampler
